+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `list` commands
+  - `roles [-d DIR] [--basemodulepath PATH]` - prints each role class & path
+  - `collectors [-i DIR] [-s DIR]` - lists collector sessions
+  - `facts [-i DIR] [-x GLOB] [selection flags]` - prints all reported facts+leaves
+  - `nodes [-i DIR] [selection flags]` - reports `all-active-nodes` inventory
+
 ## [0.4.0] - 2026-09-08
 
 ### Added